Create a proper io.buoyant.router.context package
More code reuse, too
