+++ router/core/src/main/scala/io/buoyant/router/PerDstPathFilter.scala
@@ -4,6 +4,7 @@ import com.twitter.finagle.{Path, Filter, Service, SimpleFilter}
 import com.twitter.finagle.buoyant.Dst
 import com.twitter.finagle.service.StatsFilter
 import com.twitter.util.{Future, Memoize}
+import io.buoyant.router.context.DstPathCtx
 
 class PerDstPathFilter[Req, Rsp](mk: Path => Filter[Req, Rsp, Req, Rsp])
   extends SimpleFilter[Req, Rsp] {
@@ -11,7 +12,7 @@ class PerDstPathFilter[Req, Rsp](mk: Path => Filter[Req, Rsp, Req, Rsp])
   private[this] val getFilter = Memoize(mk)
 
   def apply(req: Req, service: Service[Req, Rsp]): Future[Rsp] =
-    ctx.DstPath.current match {
+    DstPathCtx.current match {
       case None | Some(Dst.Path(Path.empty, _, _)) =>
         service(req)
