+++ router/core/src/main/scala/io/buoyant/router/context/dst.scala
@@ -1,0 +1,7 @@
+package io.buoyant.router.context
+
+import com.twitter.finagle.buoyant.Dst
+
+object DstPathCtx extends LocalKey[Dst.Path]("Dst.Path")
+
+object DstBoundCtx extends LocalKey[Dst.Bound]("Dst.Bound")
